# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Everruns API",
    "description": "API for managing AI agents, sessions, messages, and events",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "https://app.everruns.com/api",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/v1/mcp-servers": {
      "get": {
        "tags": [
          "mcp-servers"
        ],
        "summary": "GET /v1/mcp-servers - List all MCP servers",
        "operationId": "list_mcp_servers",
        "responses": {
          "200": {
            "description": "List of MCP servers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_McpServer"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "mcp-servers"
        ],
        "summary": "POST /v1/mcp-servers - Create a new MCP server",
        "operationId": "create_mcp_server",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMcpServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "MCP server created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or duplicate name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mcp-servers/{server_id}": {
      "get": {
        "tags": [
          "mcp-servers"
        ],
        "summary": "GET /v1/mcp-servers/{server_id} - Get MCP server by ID",
        "operationId": "get_mcp_server",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "description": "MCP server ID (prefixed, e.g., mcp_...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MCP server found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid server ID"
          },
          "404": {
            "description": "MCP server not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "mcp-servers"
        ],
        "summary": "DELETE /v1/mcp-servers/{server_id} - Delete MCP server",
        "operationId": "delete_mcp_server",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "description": "MCP server ID (prefixed, e.g., mcp_...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "MCP server deleted successfully"
          },
          "400": {
            "description": "Invalid server ID"
          },
          "404": {
            "description": "MCP server not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": [
          "mcp-servers"
        ],
        "summary": "PATCH /v1/mcp-servers/{server_id} - Update MCP server",
        "operationId": "update_mcp_server",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "description": "MCP server ID (prefixed, e.g., mcp_...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMcpServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "MCP server updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid server ID or input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "MCP server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMcpServerRequest": {
        "type": "object",
        "description": "Request to create a new MCP server",
        "required": [
          "name",
          "url"
        ],
        "properties": {
          "api_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "API key for authentication (optional)."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "A human-readable description of what the MCP server provides.",
            "example": "Atlassian MCP Server for Jira and Confluence"
          },
          "headers": {
            "type": [
              "object",
              "null"
            ],
            "description": "Additional HTTP headers for authentication.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": "string",
            "description": "The name of the MCP server. Must be unique.",
            "example": "atlassian-mcp-server"
          },
          "transport_type": {
            "$ref": "#/components/schemas/McpServerTransportType",
            "description": "Transport type. Currently only \"http\" is supported."
          },
          "url": {
            "type": "string",
            "description": "The URL of the MCP server endpoint.",
            "example": "https://mcp.atlassian.com/v1/mcp"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response for API endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message describing what went wrong."
          }
        }
      },
...
      "ListResponse_McpServer": {
        "type": "object",
        "description": "Response wrapper for list endpoints.\nAll list endpoints return responses wrapped in a `data` field.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "MCP Server configuration.\nRepresents a remote MCP server that can provide tools and resources.",
              "required": [
                "id",
                "name",
                "url",
                "transport_type",
                "status",
                "api_key_set",
                "created_at",
                "updated_at"
              ],
              "properties": {
                "api_key_set": {
                  "type": "boolean",
                  "description": "Whether an API key has been configured."
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Timestamp when the MCP server was created."
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Human-readable description of the MCP server.",
                  "example": "Atlassian MCP Server for Jira and Confluence"
                },
                "headers": {
                  "type": "object",
                  "description": "Additional HTTP headers for authentication.\nKeys are header names, values are header values.",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "propertyNames": {
                    "type": "string"
                  }
                },
                "id": {
                  "type": "string",
                  "description": "Unique identifier for the MCP server.",
                  "example": "mcp_01933b5a00007000800000000000001"
                },
                "name": {
                  "type": "string",
                  "description": "Display name of the MCP server.",
                  "example": "atlassian-mcp-server"
                },
                "status": {
                  "$ref": "#/components/schemas/McpServerStatus",
                  "description": "Current lifecycle status of the MCP server."
                },
                "transport_type": {
                  "$ref": "#/components/schemas/McpServerTransportType",
                  "description": "Transport type (currently only HTTP supported)."
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Timestamp when the MCP server was last updated."
                },
                "url": {
                  "type": "string",
                  "description": "URL of the MCP server endpoint.",
                  "example": "https://mcp.atlassian.com/v1/mcp"
                }
              }
            },
            "description": "Array of items returned by the list operation."
          }
        }
      },
...
      "McpServer": {
        "type": "object",
        "description": "MCP Server configuration.\nRepresents a remote MCP server that can provide tools and resources.",
        "required": [
          "id",
          "name",
          "url",
          "transport_type",
          "status",
          "api_key_set",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "api_key_set": {
            "type": "boolean",
            "description": "Whether an API key has been configured."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the MCP server was created."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable description of the MCP server.",
            "example": "Atlassian MCP Server for Jira and Confluence"
          },
          "headers": {
            "type": "object",
            "description": "Additional HTTP headers for authentication.\nKeys are header names, values are header values.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for the MCP server.",
            "example": "mcp_01933b5a00007000800000000000001"
          },
          "name": {
            "type": "string",
            "description": "Display name of the MCP server.",
            "example": "atlassian-mcp-server"
          },
          "status": {
            "$ref": "#/components/schemas/McpServerStatus",
            "description": "Current lifecycle status of the MCP server."
          },
          "transport_type": {
            "$ref": "#/components/schemas/McpServerTransportType",
            "description": "Transport type (currently only HTTP supported)."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the MCP server was last updated."
          },
          "url": {
            "type": "string",
            "description": "URL of the MCP server endpoint.",
            "example": "https://mcp.atlassian.com/v1/mcp"
          }
        }
      },
      "McpServerStatus": {
        "type": "string",
        "description": "MCP Server lifecycle status.\n- `active`: Server is available for use\n- `disabled`: Server is disabled and not used",
        "enum": [
          "active",
          "disabled"
        ]
      },
      "McpServerTransportType": {
        "type": "string",
        "description": "MCP Server transport type.\nCurrently only HTTP is supported.",
        "enum": [
          "http"
        ]
      },
...
      "UpdateMcpServerRequest": {
        "type": "object",
        "description": "Request to update an MCP server. Only provided fields will be updated.",
        "properties": {
          "api_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "API key for authentication. Set to update."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "A human-readable description of what the MCP server provides.",
            "example": "Updated description"
          },
          "headers": {
            "type": [
              "object",
              "null"
            ],
            "description": "Additional HTTP headers for authentication.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "The name of the MCP server.",
            "example": "updated-mcp-server"
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/McpServerStatus",
                "description": "The status of the MCP server. Set to \"disabled\" to disable."
              }
            ]
          },
          "transport_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/McpServerTransportType",
                "description": "Transport type."
              }
            ]
          },
          "url": {
            "type": [
              "string",
              "null"
            ],
            "description": "The URL of the MCP server endpoint.",
            "example": "https://mcp.example.com/v1/mcp"
          }
        }
      },
...
    }
  },
...
}